Add VVV2.0 Compatibility with host lists
In vvv2.0, the vvv-hosts are now saved in provisions/vvv-hosts file.
Also the sites now reside in public_html directory.

So we change the vvv_dash/hosts.php file to account for the new
hosts location. Additionally since new lines could be added, we
also pass the value through a trim.

See #52